PR feedback + test app checks wet and dry strokes
