+++ packages/live-share-inking/src/canvas/LaserPointerCanvas.ts
@@ -3,7 +3,7 @@
  * Licensed under the Microsoft Live Share SDK License.
  */
 
-import { IPointerPoint } from "../core/Geometry";
+import { IPointerPoint } from "../core";
 import { WetCanvas } from "./DryWetCanvas";
 
 /**
