+++ packages/live-share-inking/src/input/index.ts
@@ -3,8 +3,8 @@
  * Licensed under the Microsoft Live Share SDK License.
  */
 
-export * from "./InputProvider";
-export * from "./PointerInputProvider";
 export * from "./InputFilter";
+export * from "./InputProvider";
 export * from "./JitterFilter";
+export * from "./PointerInputProvider";
