--- samples/03.live-inking/.npmrc
@@ -1,2 +1,0 @@
-@ms:registry=https://office.pkgs.visualstudio.com/_packaging/Ink/npm/registry/
-always-auth=true
