+++ samples/31.inking-tests/src/index.html
@@ -24,6 +24,7 @@
             <button type="button" id="btnStartTest">Start test</button>
         </div>
     </div>
+    <div id="testResultsHost"></div>
 </body>
 
 </html>
